requirements.txt
pytest==7.4.3
requests==2.31.0
jsonschema==4.19.2
pydantic==2.5.0
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
allure-pytest==2.13.2
python-dotenv==1.0.0
faker==20.1.0
loguru==0.7.2
retry==0.9.2